requirements.txt
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
Pillow>=8.3.0
scipy>=1.7.0
tqdm>=4.62.0
configparser>=5.0.0

